#1 refactor : 회원가입 기능 코드 중 Repository에 대한 의존성이 발생하여 헥사고날 아키텍처 원칙에 따라 의존이 역전되도록 어댑터를 활용해 구현체를 구성하고 데이터베이스에 접근할 수 있도록 수정하였습니다.
