+++ src/main/java/com/example/new_instagram_server/user/application/port/service/UserRegisterService.java
@@ -1,11 +1,11 @@
-package com.example.new_instagram_server.user.port.service;
+package com.example.new_instagram_server.user.application.port.service;
 
 import com.example.new_instagram_server.user.adapter.in.UserRegisterRequestDto;
 import com.example.new_instagram_server.user.adapter.out.UserRegisterResponseDto;
+import com.example.new_instagram_server.user.application.port.in.UserRegisterUseCase;
 import com.example.new_instagram_server.user.domain.User;
 import com.example.new_instagram_server.user.domain.UserRoleType;
-import com.example.new_instagram_server.user.port.in.UserRegisterUseCase;
-import com.example.new_instagram_server.user.port.out.UserRepository;
+import com.example.new_instagram_server.user.application.port.out.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserRegisterService implements UserRegisterUseCase {
-    private final UserRepository userRepository; // 데이터베이스를 상호작용하는 인터페이스를 의존
+    private final UserRepository<User> userRepository; // 데이터베이스를 상호작용하는 인터페이스를 의존
 
     // 회원가입 로직 설계
     @Override
